Added simple animation manager to M2s
This replaces the code that directly manipulated animation tracks
in the M2 class itself.

The animation manager should make it easier to manipulate animations
from scopes external to the M2, such as: switching animations, syncing
animations with movement, etc.

Additionally, rather than establishing separate mixers for every
animation clip in the M2, a single mixer is now established. Various
function calls on the manager use actions to manipulate the state of
the mixer.