Made WMO doodad loading group-specific
Rather than queuing for load all WMO doodads at once, and placing
them as children of the WMO root, WMO doodads are now queued for
loading on a per-WMO-group basis, and are added as children of the
WMO group.

This both improves the chances that outdoor WMO groups' doodads
will have loaded by the time they are visible, and makes it easier
to do portal visibility culling in the future.