Added WMO and WMO group blueprint classes
In order to allow for cache cleanup and proper disposal of WMOs
and WMO groups, loading and unloading has been moved to blueprint
classes similar to the M2Blueprint added in a previous commit.

Additionally, logic that was related to group material creation
now happens within the WMOGroup class. This necessitated adding a
reference to the WMO root to the WMOGroup--which seems reasonable,
given that WMO groups can't really exist independent of a root.