Fixed fog issue for WMOs with transparent textures
Some textures used for WMOs are rendered as opaque (blending mode 0),
even though the texture itself has alpha transparencies. This caused
issues for the fog logic in the shader, manifested as surfaces that
would glow in the distance, even when everything else at the same
distance was obscured by fog (ie past end of fog distance).

This is now fixed by adjusting sampled pixel values to have an alpha
of 1.0 when in blending mode 0, prior to application of fog.

Example texture: DUNGEONS/TEXTURES/WALLS/MM_STRMWND_WALL_04.BLP
(used on the smithy WMO in Darkshire)