Fix render count warning
It turns out the render count WebGL spam that appears in the browser
console is because the 'root' mesh in every M2 has no material associated
with it, and thus isn't renderable.

At some point, we should really revisit how the geometry and meshes
in M2s are handled. There probably is no purpose for the root mesh.