Split WMODefinition out from WMOBlueprint
WMODefinition is used to set up transferable arrays of data on
worker threads.

WMOBlueprint is used to set up various objects (including rendering
related objects) on the main thread.

As of this change, blueprints are no longer confusingly used in two
separate contexts. Additionally, the worker is back to being
relatively small, and with far fewer dependencies.