Added camera location mechanism
Portal culling originates from the camera's location, which
necessitates determining that location prior to the start of
portal traversal.

The camera can either be located outside of all WMOs, inside a
specific WMO group flagged as exterior, or inside a specific WMO
group flagged as interior.

In order to correctly determine the location, even when dealing
with overlapping WMO groups, the WMO group BSP trees are loaded
and queried for the camera's location in local space. WMO and
WMO group bounding boxes are checked to minimize the number of
useless BSP tree queries.