Initial implementation of VisibilityManager
The visibility manager is responsible for walking the scene and
flagging potentially visible objects as visible.

At the start of an update, the visibility manager flags all content
in the scene is not visible.

It then traverses WMO portals, creates frustums matching the portals,
and ensures objects that objects intersecting with these frustums
are marked as visible.